pyproject.toml
[project]
name = "kuhaku-agent-cli"
version = "0.1.0"
description = "Bridge messaging surfaces (Slack first) to Claude Managed Agents"
readme = "README.md"
requires-python = ">=3.12"
license = "Apache-2.0"
license-files = ["LICENSE", "NOTICE"]
dependencies = [
    "anthropic>=0.97.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "slack-bolt>=1.28.0",
    "slack-sdk>=3.41.0",
    "typer>=0.15.0",
    "rich>=13.9.0",
]

[project.scripts]
kuhaku-agent = "kuhaku_agent.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kuhaku_agent"]
